requirement.txt
attrs==19.1.0
Automat==0.7.0
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.5.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.6.1
cssselect==1.0.3
decorator==4.0.10
dict2xml==1.5
Django==2.0.7
django-celery==3.2.2
django-cors-headers==2.2.0
django-crontab==0.7.1
django-cursor-pagination==0.1.2
django-filter==2.0.0
django-redis==4.9.0
django-redis-cache==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
drf-dynamic-fields==0.3.0
drf-friendly-errors==0.13
Faker==2.0.3
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
ipdb==0.10.1
IPy==1.0
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.3.17
itypes==1.1.0
jsonschema==3.0.1
kombu==4.5.0
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.17.0
mysqlclient==1.3.13
numpy==1.16.3
openapi-codec==1.3.2
packaging==16.8
parsel==1.5.1
pexpect==4.2.1
pickleshare==0.7.4
pika==1.0.1
Pillow==5.3.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.1.3
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyotp==2.2.6
pyparsing==2.2.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2018.4
queuelib==1.5.0
redis==3.2.1
requests==2.20.1
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
tornado==5.1.1
traitlets==4.3.1
uritemplate==3.0.0
urllib3==1.24.2
uWSGI==2.0.18
wordcloud==1.6.0
